Move ipam/address to net/address
