+++ ipam/ring/entry.go
@@ -4,7 +4,7 @@ import (
 	"sort"
 
 	"github.com/weaveworks/weave/common"
-	"github.com/weaveworks/weave/ipam/address"
+	"github.com/weaveworks/weave/net/address"
 	"github.com/weaveworks/weave/router"
 )
