+++ ipam/space/space.go
@@ -6,7 +6,7 @@ import (
 	"sort"
 
 	"github.com/weaveworks/weave/common"
-	"github.com/weaveworks/weave/ipam/address"
+	"github.com/weaveworks/weave/net/address"
 )
 
 type Space struct {
